feat: add athorization layer to handlers

+++ controllers/folder.controller.ts
@@ -5,11 +5,14 @@ import getDBClient from '../db/client';
 import { createProtectedHandler } from '@/utils/createHandler';
 import { z } from 'zod';
 import { singleOrThrow } from '@/db/utils';
+import { defineAbilitiesFor } from '@/access-control/user.access';
+import { subject } from '@casl/ability';
 
 const db = getDBClient();
 
 export const getFolders = createProtectedHandler(
   z.object({}),
+  () => true,
   async (req, res) => {
     try {
       // const filterOptions = req.query.userId ? { author: req.query.userId } : {};
@@ -24,8 +27,14 @@ export const getFolders = createProtectedHandler(
 
 export const getFolder = createProtectedHandler(
   z.object({ params: z.object({ id: z.string() }) }),
+  (ability, req) => ability.can('read', subject('Folder', req.body)),
   async (req, res) => {
     try {
+
+      const ability = defineAbilitiesFor(req.user)
+      const hasAccess = ability.can('create', subject('Folder', req.body))
+      if (!hasAccess) throw Error('Unauthorized');
+
       const { id } = req.params;
       const folder = await db.select().from(Folder).where(eq(Folder.id, id)).then(singleOrThrow);
       // Folder.findById(id).populate('author').populate('lists');
@@ -38,8 +47,10 @@ export const getFolder = createProtectedHandler(
 
 export const createFolder = createProtectedHandler(
   z.object({ params: z.object({ id: z.string() }), body: folderCreateSchema }),
+  (ability, req) => ability.can('create', subject('Folder', req.body)),
   async (req, res) => {
     try {
+
       const folder = await db.insert(Folder).values(req.body).returning().then(singleOrThrow);
       res.status(200).json(folder);
     } catch (error) {
@@ -50,10 +61,14 @@ export const createFolder = createProtectedHandler(
 
 export const updateFolder = createProtectedHandler(
   z.object({ params: z.object({ id: z.string() }), body: folderUpdateSchema }),
+  async (ability , req) => {
+    const { id } = req.params;
+    const currentFolder = await db.select().from(Folder).where(eq(Folder.id, id)).then(singleOrThrow);
+    return ability.can('update', subject('Folder', { ...req.body, id: currentFolder.id }));
+  }, 
   async (req, res) => {
     try {
       const { id } = req.params;
-
       const folder = await db.update(Folder).set(req.body).where(eq(Folder.id, id)).returning().then(singleOrThrow);
 
       if (!folder) {
@@ -69,10 +84,19 @@ export const updateFolder = createProtectedHandler(
 );
 
 export const deleteFolder = createProtectedHandler(
-  z.object({ params: z.object({ id: z.string() }) }),
+  z.object({ params: z.object({ id: z.string() }) , body: folderCreateSchema }),
+  async (ability, req) => {
+    const { id } = req.params;
+    const currentFolder = await db.select().from(Folder).where(eq(Folder.id, id)).then(singleOrThrow);
+    return ability.can('delete', subject('Folder', { ...req.body, id: currentFolder.id }));
+  },
   async (req, res) => {
     try {
       const { id } = req.params;
+
+      const ability = defineAbilitiesFor(req.user)
+      const hasAccess = ability.can('delete', subject('Folder', req.body));
+      if (!hasAccess) throw Error('Unauthorized');
 
       const folder = await db.delete(Folder).where(eq(Folder.id, id)).returning().then(singleOrThrow);
