+++ controllers/list.controller.ts
@@ -13,6 +13,7 @@ const db = getDBClient();
 
 export const getLists = createProtectedHandler(
   z.object({}),
+  () => true,
   async (req, res) => {
     try {
       // @ts-ignore
@@ -24,18 +25,26 @@ export const getLists = createProtectedHandler(
     }
   });
 
-export const getList = createProtectedHandler(z.object({ params: z.object({ id: z.string() }) }), async (req, res) => {
-  try {
-    const { id } = req.params;
-    const list = await db.select().from(List).where(eq(List.id, id)).then(singleOrThrow);
-    // List.findById(id).populate('author').populate('tasks.task');
-    res.status(200).json(list);
-  } catch (error) {
-    res.status(500).json({ message: error instanceof Error ? error.message : "Internal Server Error: " + error });
-  }
-});
+export const getList = createProtectedHandler(z.object({ params: z.object({ id: z.string() }) }),
+  (ability, req) => ability.can('read', subject('List', req.body)),
+  async (req, res) => {
+    try {
+      const ability = defineAbilitiesFor(req.user)
+      const hasAccess = ability.can('read', subject('List', req.body));
+      if (!hasAccess) throw Error('Unauthorized');
 
-export const createList = createProtectedHandler(z.object({ params: z.object({ id: z.string() }), body: listCreateSchema }), async (req, res) => {
+      const { id } = req.params;
+      const list = await db.select().from(List).where(eq(List.id, id)).then(singleOrThrow);
+      // List.findById(id).populate('author').populate('tasks.task');
+      res.status(200).json(list);
+    } catch (error) {
+      res.status(500).json({ message: error instanceof Error ? error.message : "Internal Server Error: " + error });
+    }
+  });
+
+export const createList = createProtectedHandler(z.object({ params: z.object({ id: z.string() }), body: listCreateSchema }),
+(ability, req) => ability.can('create', subject('List', req.body)),
+async (req, res) => {
   try {
 
     const ability = defineAbilitiesFor(req.user)
@@ -49,12 +58,22 @@ export const createList = createProtectedHandler(z.object({ params: z.object({ i
   }
 });
 
-export const updateList = createProtectedHandler(z.object({ body: listUpdateSchema, params: z.object({ id: z.string() }) }), async (req, res) => {
+export const updateList = createProtectedHandler(z.object({ body: listUpdateSchema, params: z.object({ id: z.string() }) }),
+async (ability , req) => {
+  const { id } = req.params;
+  const currentList = await db.select().from(List).where(eq(List.id, id)).then(singleOrThrow);
+  return ability.can('update', subject('List', { ...req.body, id: currentList.id }));
+}, 
+async (req, res) => {
   try {
-    const ability = defineAbilitiesFor(req.user as UserSelect)
-    const hasAccess = ability.can('update', subject('List', req.body));
-    if (!hasAccess) throw Error('Unauthorized');
+
     const { id } = req.params;
+    const currentList = await db.select().from(List).where(eq(List.id, id)).then(singleOrThrow);
+
+    const ability = defineAbilitiesFor(req.user)
+    const hasAccess = ability.can('update', subject('List', { ...req.body, authorId: currentList.authorId }));
+    if (!hasAccess) throw Error('Unauthorized');
+
     const list = await db.update(List).set(req.body).where(eq(List.id, id)).returning().then(singleOrThrow);
     // List.findByIdAndUpdate(id, req.body);
 
@@ -69,8 +88,20 @@ export const updateList = createProtectedHandler(z.object({ body: listUpdateSche
   }
 });
 
-export const deleteList = createProtectedHandler(z.object({ params: z.object({ id: z.string() }) }), async (req, res) => {
+export const deleteList = createProtectedHandler(z.object({ params: z.object({ id: z.string() }) }),
+
+async (ability, req) => {
+  const { id } = req.params;
+  const currentList = await db.select().from(List).where(eq(List.id, id)).then(singleOrThrow);
+  return ability.can('delete', subject('List', { ...req.body, id: currentList.id }));
+},
+async (req, res) => {
   try {
+
+    const ability = defineAbilitiesFor(req.user)
+    const hasAccess = ability.can('delete', subject('List', req.body));
+    if (!hasAccess) throw Error('Unauthorized');
+
     const { id } = req.params;
 
     const list = await db.delete(List).where(eq(List.id, id)).returning().then(singleOrThrow);
